pi/backend/pyproject.toml
[project]
name = "smartserow-backend"
version = "0.1.0"
description = "GPS and Arduino telemetry service for Smart Serow"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0",
    "gpsdclient>=1.3",
    "pyserial>=3.5",
]

[project.optional-dependencies]
dev = [
    "ruff",
]

[project.scripts]
smartserow-backend = "main:main"

[tool.ruff]
line-length = 100
